% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsn_functions.R
\name{get_dsn}
\alias{get_dsn}
\title{Get DSNs and Default DSN}
\usage{
get_dsn()
}
\value{
Returns a printed list of your current set DSNs within in this session.
}
\description{
See a printed list of your DSNs in this session and corresponding aliases.
}
\examples{
get_dsn()
}
