% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{read_query}
\alias{read_query}
\title{Read a SQL Query From a Text File}
\usage{
read_query(file)
}
\arguments{
\item{file}{The directory/filename of the text file.}
}
\value{
A character string from the text file.
}
\description{
This is function allows the user to read a query from a text file. No validation
of that file is conducted. In other words, this function could read any text file.
For functions that use multiple temp table steps. Separate each query by '---step---'
so that each step can be idenfitied. Don't do this for CTE expressions.
}
\examples{
\dontrun{read_query("my_file.txt")}
}
