% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsn_functions.R
\name{save_dsn}
\alias{save_dsn}
\title{Manually Save Your DSN List to .Renviron}
\usage{
save_dsn()
}
\value{
Opens the .Renviron edit window.
}
\description{
Allows you to modify your .Renviron to permanently save your session DSNs.
}
\examples{
\dontrun{save_dsn()}
}
