% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{sql_query}
\alias{sql_query}
\title{Execute a single query}
\usage{
sql_query(
  query,
  dsn_alias = names(options()$ez.dsn.default),
  close = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Execute a single query
}
