% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsn_functions.R
\name{reset_dsn}
\alias{reset_dsn}
\title{Clear Your DSN List}
\usage{
reset_dsn()
}
\value{
Clears the DSNs out of your current session.
}
\description{
Clears your current DSN list for this session.
}
\examples{
reset_dsn() # Be careful with this!

}
