% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{sql_query}
\alias{sql_query}
\title{Execute a single query}
\usage{
sql_query(
  query,
  dsn_alias = names(options()$ez.dsn.default),
  close = TRUE,
  ...
)
}
\arguments{
\item{query}{The query from which you would like to get data.}

\item{close}{Close and destroy the connection once the query is complete?}

\item{...}{Additional arguments passed to odbc::dbGetQuery()}

\item{dsn}{The alias of your dsn. Uses default dsn if not specified.}
}
\value{
Presuming your query is written properly. Returns a data frame of data
  (e.g. isn't inserting data into a temp table)
}
\description{
Execute a single query
}
\examples{
\dontrun{sql_query("select foo from bar where beans = 'beans'")}
}
